Explicitly define all supported python environment
With only 3.5, travis-ci works, because 2.7, 3.3 and 3.4 interpreters
are present by default,  but it might break randomly at some point.

Definining all the versions explicitly, and using the tox-travis plugin
is more robust.

Also, 2.6 was removed from the "officially supported" python versions,
as the tests are not run (and don't pass).